Remove Component namespace on annotation
