+++ src/Bundle/Routing/RouteAttributesFactory.php
@@ -13,8 +13,8 @@ declare(strict_types=1);
 
 namespace Sylius\Bundle\ResourceBundle\Routing;
 
-use Sylius\Component\Resource\Annotation\SyliusRoute;
 use Sylius\Component\Resource\Reflection\ClassReflection;
+use Sylius\Resource\Annotation\SyliusRoute;
 use Symfony\Component\Routing\Route;
 use Symfony\Component\Routing\RouteCollection;
 use Webmozart\Assert\Assert;
