+++ tests/Application/src/Entity/CrudRoutes/BookWithRedirect.php
@@ -15,7 +15,7 @@ namespace App\Entity\CrudRoutes;
 
 use App\Entity\Book;
 use JMS\Serializer\Annotation as Serializer;
-use Sylius\Component\Resource\Annotation\SyliusCrudRoutes;
+use Sylius\Resource\Annotation\SyliusCrudRoutes;
 
 /**
  * @Serializer\ExclusionPolicy("all")
