+++ tests/Application/src/Entity/Route/ShowBook.php
@@ -15,7 +15,7 @@ namespace App\Entity\Route;
 
 use App\Entity\Book;
 use JMS\Serializer\Annotation as Serializer;
-use Sylius\Component\Resource\Annotation\SyliusRoute;
+use Sylius\Resource\Annotation\SyliusRoute;
 
 /**
  * @Serializer\ExclusionPolicy("all")
