+++ tests/Application/src/Entity/Route/UpdateBookWithCustomEventName.php
@@ -14,7 +14,7 @@ declare(strict_types=1);
 namespace App\Entity\Route;
 
 use App\Entity\Book;
-use Sylius\Component\Resource\Annotation\SyliusRoute;
+use Sylius\Resource\Annotation\SyliusRoute;
 
 #[SyliusRoute(
     name: 'update_book_with_custom_event_name',
